Added header files

+++ templates/data-structures/policybasedDS.h
@@ -4,12 +4,12 @@ template<class T>
 using Tree = tree<T, null_type, less<T>, rb_tree_tag,
     tree_order_statistics_node_update>;
 struct chash { /// use most bits rather than just the lowest ones
-	const uint64_t C = ll(2e18*PI)+71; // large odd number
-	const int RANDOM = rng();
+	const uint64_t C = ll(2e18 * PI) + 71; // large odd number
+	const int RANDOM = (int)rng();
 	ll operator()(ll x) const { /// https://gcc.gnu.org/onlinedocs/gcc/Other-Builtins.html
-		return __builtin_bswap64((x^RANDOM)*C); }
+		return __builtin_bswap64((x ^ RANDOM) * C); }
 };
-template<class K,class V> using um = unordered_map<K,V,chash>;
-template<class K,class V> using ht = gp_hash_table<K,V,chash>;
-template<class K,class V> V get(ht<K,V>& u, K x) {
+template <class K, class V> using um = unordered_map<K, V, chash>;
+template <class K, class V> using ht = gp_hash_table<K, V, chash>;
+template <class K, class V> V get(ht<K, V>& u, K x) {
 	auto it = u.find(x); return it == end(u) ? 0 : it->s; }
